% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/has-attributes.R
\name{has_any_attributes}
\alias{has_any_attributes}
\title{Does the input have any attributes?}
\usage{
has_any_attributes(x, .xname = get_name_in_parent(x))
}
\arguments{
\item{x}{Input to check.}

\item{.xname}{Not intended to be used directly.}
}
\value{
\code{has_any_attributes} returns \code{TRUE} if \code{attributes(x)}
has length greater than zero. \code{has_attributes} returns a logical vector
that is \code{TRUE} whenever the specified attribute is not \code{NULL}.

\code{assert_has_all_attributes} and \code{assert_has_all_attributes} return
nothing but throw an error if \code{has_attributes} is not \code{TRUE}.
}
\description{
Checks to see if the input has any attributes.
}
\examples{
has_any_attributes(matrix(1:12, nrow = 3))
}

